sonar
git-svn-id: http://josm.openstreetmap.de/svn/trunk@11048 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/tools/date/DateUtils.java
@@ -97,7 +97,7 @@ public final class DateUtils {
             );
             if (str.length() == 22 || str.length() == 25) {
                 final int plusHr = parsePart2(str, 20);
-                final int mul = str.charAt(19) == '+' ? -1 : 1;
+                final long mul = str.charAt(19) == '+' ? -1 : 1;
                 return local.plusHours(plusHr * mul).toInstant().toEpochMilli();
             }
             return local.toInstant().toEpochMilli();
@@ -119,7 +119,7 @@ public final class DateUtils {
             );
             if (str.length() == 29) {
                 final int plusHr = parsePart2(str, 24);
-                final int mul = str.charAt(23) == '+' ? -1 : 1;
+                final long mul = str.charAt(23) == '+' ? -1 : 1;
                 return local.plusHours(plusHr * mul).toInstant().toEpochMilli();
             }
             return local.toInstant().toEpochMilli();
